Module google_cloudmonitoring2_beta2::api

Structs

CloudMonitoring
Central instance to access all CloudMonitoring related resource activities

DeleteMetricDescriptorResponse
The response of cloudmonitoring.metricDescriptors.delete.

ListMetricDescriptorsRequest
The request of cloudmonitoring.metricDescriptors.list.

ListMetricDescriptorsResponse
The response of cloudmonitoring.metricDescriptors.list.

ListTimeseriesDescriptorsRequest
The request of cloudmonitoring.timeseriesDescriptors.list

ListTimeseriesDescriptorsResponse
The response of cloudmonitoring.timeseriesDescriptors.list

ListTimeseriesRequest
The request of cloudmonitoring.timeseries.list

ListTimeseriesResponse
The response of cloudmonitoring.timeseries.list

MetricDescriptor
A metricDescriptor defines the name, label keys, and data type of a particular metric.

MetricDescriptorCreateCall
Create a new metric.

MetricDescriptorDeleteCall
Delete an existing metric.

MetricDescriptorLabelDescriptor
A label in a metric is a description of this metric, including the key of this description (what the description is), and the value for this description.

MetricDescriptorListCall
List metric descriptors that match the query. If the query is not set, then all of the metric descriptors will be returned. Large responses will be paginated, use the nextPageToken returned in the response to request subsequent pages of results by setting the pageToken query parameter to the value of the nextPageToken.

MetricDescriptorMethods
A builder providing access to all methods supported on metricDescriptor resources. It is not used directly, but through the CloudMonitoring hub.

MetricDescriptorTypeDescriptor
A type in a metric contains information about how the metric is collected and what its data points look like.

Point
Point is a single point in a time series. It consists of a start time, an end time, and a value.

PointDistribution
Distribution data point value type. When writing distribution points, try to be consistent with the boundaries of your buckets. If you must modify the bucket boundaries, then do so by merging, partitioning, or appending rather than skewing them.

PointDistributionBucket
The histogram's bucket. Buckets that form the histogram of a distribution value. If the upper bound of a bucket, say U1, does not equal the lower bound of the next bucket, say L2, this means that there is no event in [U1, L2).

PointDistributionOverflowBucket
The overflow bucket is a special bucket that does not have the upperBound field; it includes all of the events that are no less than its lower bound.

PointDistributionUnderflowBucket
The underflow bucket is a special bucket that does not have the lowerBound field; it includes all of the events that are less than its upper bound.

Timeseries
The monitoring data is organized as metrics and stored as data points that are recorded over time. Each data point represents information like the CPU utilization of your virtual machine. A historical record of these data points is called a time series.

TimeseriesDescriptor
TimeseriesDescriptor identifies a single time series.

TimeseriesDescriptorListCall
List the descriptors of the time series that match the metric and labels values and that have data points in the interval. Large responses are paginated; use the nextPageToken returned in the response to request subsequent pages of results by setting the pageToken query parameter to the value of the nextPageToken.

TimeseriesDescriptorMethods
A builder providing access to all methods supported on timeseriesDescriptor resources. It is not used directly, but through the CloudMonitoring hub.

TimeseriesPoint
When writing time series, TimeseriesPoint should be used instead of Timeseries, to enforce single point for each time series in the timeseries.write request.

TimeseryListCall
List the data points of the time series that match the metric and labels values and that have data points in the interval. Large responses are paginated; use the nextPageToken returned in the response to request subsequent pages of results by setting the pageToken query parameter to the value of the nextPageToken.

TimeseryMethods
A builder providing access to all methods supported on timesery resources. It is not used directly, but through the CloudMonitoring hub.

TimeseryWriteCall
Put data points to one or more time series for one or more metrics. If a time series does not exist, a new time series will be created. It is not allowed to write a time series point that is older than the existing youngest point of that time series. Points that are older than the existing youngest point of that time series will be discarded silently. Therefore, users should make sure that points of a time series are written sequentially in the order of their end time.

WriteTimeseriesRequest
The request of cloudmonitoring.timeseries.write

WriteTimeseriesResponse
The response of cloudmonitoring.timeseries.write

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.