Struct google_cloudmonitoring2_beta2::PointDistribution

pub struct PointDistribution {
    pub buckets: Option<Vec<PointDistributionBucket>>,
    pub underflow_bucket: Option<PointDistributionUnderflowBucket>,
    pub overflow_bucket: Option<PointDistributionOverflowBucket>,
}

Distribution data point value type. When writing distribution points, try to be consistent with the boundaries of your buckets. If you must modify the bucket boundaries, then do so by merging, partitioning, or appending rather than skewing them.

This type is not used in any activity, and only used as part of another schema.

Fields
buckets: Option<Vec<PointDistributionBucket>>

The finite buckets.

underflow_bucket: Option<PointDistributionUnderflowBucket>

The underflow bucket.

overflow_bucket: Option<PointDistributionOverflowBucket>

The overflow bucket.