Struct google_cloudmonitoring2_beta2::TimeseriesDescriptor

pub struct TimeseriesDescriptor {
    pub project: Option<String>,
    pub metric: Option<String>,
    pub labels: Option<HashMap<String, String>>,
}

TimeseriesDescriptor identifies a single time series.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list timeseries descriptors (none)

Fields
project: Option<String>

The Developers Console project number to which this time series belongs.

metric: Option<String>

The name of the metric.

labels: Option<HashMap<String, String>>

The label's name.