Struct google_cloudmonitoring2_beta2::TimeseriesDescriptorMethods

pub struct TimeseriesDescriptorMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on timeseriesDescriptor resources. It is not used directly, but through the CloudMonitoring hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_cloudmonitoring2_beta2 as cloudmonitoring2_beta2;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use cloudmonitoring2_beta2::CloudMonitoring;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = CloudMonitoring::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `list(...)`
// to build up your call.
let rb = hub.timeseries_descriptors();

Methods

impl<'a, C, A> TimeseriesDescriptorMethods<'a, C, A>

fn list(&self, request: ListTimeseriesDescriptorsRequest, project: &str, metric: &str, youngest: &str) -> TimeseriesDescriptorListCall<'a, C, A>

Create a builder to help you perform the following task:

List the descriptors of the time series that match the metric and labels values and that have data points in the interval. Large responses are paginated; use the nextPageToken returned in the response to request subsequent pages of results by setting the pageToken query parameter to the value of the nextPageToken.

Arguments

request - No description provided.

project - The project ID to which this time series belongs. The value can be the numeric project ID or string-based project name.

metric - Metric names are protocol-free URLs as listed in the Supported Metrics page. For example, compute.googleapis.com/instance/disk/read_ops_count.

youngest - End of the time interval (inclusive), which is expressed as an RFC 3339 timestamp.