Struct google_cloudmonitoring2_beta2::TimeseryListCall

pub struct TimeseryListCall<'a, C, A> where
    C: 'a,
    A: 'a,  { /* fields omitted */ }

List the data points of the time series that match the metric and labels values and that have data points in the interval. Large responses are paginated; use the nextPageToken returned in the response to request subsequent pages of results by setting the pageToken query parameter to the value of the nextPageToken.

A builder for the list method supported by a timesery resource. It is not used directly, but through a TimeseryMethods instance.

Example

Instantiate a resource method builder

use cloudmonitoring2_beta2::ListTimeseriesRequest;
 
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req = ListTimeseriesRequest::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.timeseries().list(req, "project", "metric", "youngest")
             .window("ipsum")
             .timespan("Lorem")
             .page_token("et")
             .oldest("duo")
             .add_labels("aliquyam")
             .count(-9)
             .aggregator("Lorem")
             .doit();

Methods

impl<'a, C, A> TimeseryListCall<'a, C, A> where
C: BorrowMut<Client>,
A: GetToken,

pub fn doit(self) -> Result<(Response, ListTimeseriesResponse)>

Perform the operation you have build so far.

pub fn request(
self,
new_value: ListTimeseriesRequest
) -> TimeseryListCall<'a, C, A>

Sets the request property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

pub fn project(self, new_value: &str) -> TimeseryListCall<'a, C, A>

The project ID to which this time series belongs. The value can be the numeric project ID or string-based project name.

Sets the project path property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

pub fn metric(self, new_value: &str) -> TimeseryListCall<'a, C, A>

Metric names are protocol-free URLs as listed in the Supported Metrics page. For example, compute.googleapis.com/instance/disk/read_ops_count.

Sets the metric path property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

pub fn youngest(self, new_value: &str) -> TimeseryListCall<'a, C, A>

End of the time interval (inclusive), which is expressed as an RFC 3339 timestamp.

Sets the youngest query property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

pub fn window(self, new_value: &str) -> TimeseryListCall<'a, C, A>

The sampling window. At most one data point will be returned for each window in the requested time interval. This parameter is only valid for non-cumulative metric types. Units:

m: minute

h: hour

d: day

w: week Examples: 3m, 4w. Only one unit is allowed, for example: 2w3d is not allowed; you should use 17d instead.

Sets the window query property to the given value.

pub fn timespan(self, new_value: &str) -> TimeseryListCall<'a, C, A>

Length of the time interval to query, which is an alternative way to declare the interval: (youngest - timespan, youngest]. The timespan and oldest parameters should not be used together. Units:

s: second

m: minute

h: hour

d: day

w: week Examples: 2s, 3m, 4w. Only one unit is allowed, for example: 2w3d is not allowed; you should use 17d instead.

If neither oldest nor timespan is specified, the default time interval will be (youngest - 4 hours, youngest].

Sets the timespan query property to the given value.

pub fn page_token(self, new_value: &str) -> TimeseryListCall<'a, C, A>

The pagination token, which is used to page through large result sets. Set this value to the value of the nextPageToken to retrieve the next page of results.

Sets the page token query property to the given value.

pub fn oldest(self, new_value: &str) -> TimeseryListCall<'a, C, A>

Start of the time interval (exclusive), which is expressed as an RFC 3339 timestamp. If neither oldest nor timespan is specified, the default time interval will be (youngest - 4 hours, youngest]

Sets the oldest query property to the given value.

pub fn add_labels(self, new_value: &str) -> TimeseryListCall<'a, C, A>

A collection of labels for the matching time series, which are represented as:

key==value: key equals the value

key=~value: key regex matches the value

key!=value: key does not equal the value

key!~value: key regex does not match the value For example, to list all of the time series descriptors for the region us-central1, you could specify: label=cloud.googleapis.com%2Flocation=~us-central1.*

Append the given value to the labels query property. Each appended value will retain its original ordering and be '/'-separated in the URL's parameters.

pub fn count(self, new_value: i32) -> TimeseryListCall<'a, C, A>

Maximum number of data points per page, which is used for pagination of results.

Sets the count query property to the given value.

pub fn aggregator(self, new_value: &str) -> TimeseryListCall<'a, C, A>

The aggregation function that will reduce the data points in each window to a single point. This parameter is only valid for non-cumulative metrics with a value type of INT64 or DOUBLE.

Sets the aggregator query property to the given value.

pub fn delegate(
self,
new_value: &'a mut dyn Delegate
) -> TimeseryListCall<'a, C, A>

The delegate implementation is consulted whenever there is an intermediate result, or if something goes wrong while executing the actual API request.

It should be used to handle progress information, and to implement a certain level of resilience.

Sets the delegate property to the given value.

pub fn param<T>(self, name: T, value: T) -> TimeseryListCall<'a, C, A> where
T: AsRef<str>,

Set any additional parameter of the query string used in the request. It should be used to set parameters which are not yet available through their own setters.

Please note that this method must not be used to set any of the known parameters which have their own setter method. If done anyway, the request will fail.

Additional Parameters

quotaUser (query-string) - Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters. Overrides userIp if both are provided.

oauth_token (query-string) - OAuth 2.0 token for the current user.

key (query-string) - API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

prettyPrint (query-boolean) - Returns response with indentations and line breaks.

userIp (query-string) - IP address of the site where the request originates. Use this if you want to enforce per-user limits.

fields (query-string) - Selector specifying which fields to include in a partial response.

alt (query-string) - Data format for the response.

pub fn add_scope<T, S>(self, scope: T) -> TimeseryListCall<'a, C, A> where
T: Into<Option<S>>,
S: AsRef<str>,

Identifies the authorization scope for the method you are building.

Use this method to actively specify which scope should be used, instead the default Scope variant Scope::CloudPlatform.

The scope will be added to a set of scopes. This is important as one can maintain access tokens for more than one scope. If None is specified, then all scopes will be removed and no default scope will be used either. In that case, you have to specify your API-key using the key parameter (see the param() function for details).

Usually there is more than one suitable scope to authorize an operation, some of which may encompass more rights than others. For example, for listing resources, a read-only scope will be sufficient, a read-write scope will do as well.